+++ modules/hanna-react/src/utils.ts
@@ -1,3 +1,4 @@
+export * from './utils/config';
 export * from './utils/useDidChange';
 export * from './utils/useFormatMonitor';
 export * from './utils/useGetSVGtext';